backend/requirements.txt
aioredis==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
billiard==3.6.4.0
black==22.6.0
boto3==1.24.23
botocore==1.27.23
CacheControl==0.12.11
cachetools==5.3.0
celery==5.2.1
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coolsms-python-sdk==2.0.3
cryptography==39.0.0
daphne==3.0.2
Django==4.1.4
django-ckeditor==6.4.2
django-cors-headers==3.10.0
django-filter==21.1
django-hosts==5.1
django-js-asset==2.0.0
django-safedelete==1.3.1
django-seed==0.3.1
django-storages==1.12.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.25.1
Faker==15.3.4
grpcio==1.51.1
grpcio-status==1.51.1
gunicorn==20.1.0
hiredis==2.1.0
httplib2==0.21.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
isort==5.10.1
jmespath==1.0.1
jsonschema==4.17.3
kombu==5.2.4
msgpack==1.0.4
mypy-extensions==0.4.3
numpy==1.23.5
pathspec==0.10.3
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
proto-plus==1.22.2
protobuf==4.21.12
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.7
PyYAML==6.0
redis==4.5.1
requests==2.26.0
rsa==4.9
s3transfer==0.6.0
sentry-sdk==1.6.0
service-identity==21.1.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toposort==1.9
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
watchtower==3.0.0
wcwidth==0.2.5
zope.interface==5.5.2
